fix(deploy): monetary columns bigint + pg int8 parser (Postgres overflow)
Postgres integer (4-byte, max ~2.1B) overflowed on cents values (e.g.
$514B = 51 trillion cents). SQLite int is 8-byte so it never surfaced
locally. Changed obligatedAmount/outlayAmount/femaObligated/
fedSpendingObligated/perCapita columns from integer to bigint across
all 4 entities. Added pg types.setTypeParser(INT8, Number) in main.ts
so bigint columns return JS numbers (safe: max value ~5e13 < 2^53).
Switched Postgres to synchronize:true (migrationsRun:false) because
the Init migration used pluralized table names (agencies) that don't
match the entity defaults (agency) - synchronize creates correct
schema from entities on fresh deploy.

Verified: Neon Postgres accepts the values; agencies+disaster synced.
@types/pg added for the pg import. typeorm.config.spec updated.
backend 210 tests green.

+++ backend/src/disaster/disaster-funding-record.entity.ts
@@ -18,13 +18,13 @@ export class DisasterFundingRecord {
   @Column({ type: 'varchar', length: 255 })
   stateName: string;
 
-  @Column({ type: 'integer' })
+  @Column({ type: 'bigint' })
   obligatedAmount: number;
 
   @Column()
   awardCount: number;
 
-  @Column({ type: 'integer' })
+  @Column({ type: 'bigint' })
   perCapita: number;
 
   @Column()
